requirements.txt
sounddevice
aubio
numpy
scipy
pyserial
